Improve historyScreen UI
